:hammer: Rename term-extractor to extract-terms
