+++ src/extract-terms.js
