+++ src/index.js
@@ -2,7 +2,7 @@ import TEEFTTokenize from './tokenize';
 import TEEFTFrToTagLem from './fr-to-tag-lem';
 import TEEFTStopWords from './stop-words';
 import TEEFTFilterTags from './filter-tags';
-import TEEFTExtractTerms from './term-extractor';
+import TEEFTExtractTerms from './extract-terms';
 import TEEFTSumUpFrequencies from './sum-up-freq';
 import TEEFTFilterMonoFreq from './filter-mono-freq';
 import TEEFTSpecificity from './specificity';
